requirements.txt
langchain
langchain-core
langchain-openai
langgraph
pydantic>=2.0
pypdf
python-dotenv
